Move eye_movement models into separate files
