+++ pupil_src/shared_modules/eye_movement/model/immutable_capture.py
@@ -1,0 +1,18 @@
+import typing as t
+from camera_models import Camera_Model
+from video_capture.base_backend import Base_Source
+import numpy as np
+
+
+class Immutable_Capture:
+    def __init__(self, capture: t.Type[Base_Source]):
+        self.frame_size: t.Tuple[int, int] = (
+            int(capture.frame_size[0]),
+            int(capture.frame_size[1]),
+        )
+        self.intrinsics: t.Type[Camera_Model] = capture.intrinsics
+        try:
+            self.timestamps: np.ndarray = capture.timestamps
+        except AttributeError:
+            self.timestamps: np.ndarray = np.ndarray([])
+
