+++ pupil_src/shared_modules/eye_movement_detector.py
@@ -48,485 +48,6 @@ from pyglui.cygl.utils import RGBA, draw_circle
 from pyglui.pyfontstash import fontstash
 
 
-class Time_Range():
-    def __init__(self, start_time: float, end_time: float):
-        assert start_time <= end_time
-        self.start_time = start_time
-        self.end_time = end_time
-
-    def contains(self, timestamp: float) -> bool:
-        return self.start_time <= timestamp and timestamp <= self.end_time
-
-    def intersection(self, other: 'Time_Range') -> typing.Optional['Time_Range']:
-        start_time = max(self.start_time, other.start_time)
-        end_time = min(self.end_time, other.end_time)
-        if start_time <= end_time:
-            return Time_Range(start_time=start_time, end_time=end_time)
-        else:
-            return None
-
-    def union(self, other: 'Time_Range') -> 'Time_Range':
-        """Warning: This method doesn't guarantee that the union is continuous."""
-        start_time = min(self.start_time, other.start_time)
-        end_time = max(self.end_time, other.end_time)
-        return Time_Range(start_time=start_time, end_time=end_time)
-
-    def countinuous_union(self, other: 'Time_Range') -> typing.Optional['Time_Range']:
-        if not self.intersection(other):
-            return None
-        return self.union(other)
-
-
-class Immutable_Capture:
-    def __init__(self, capture: typing.Type[vc.base_backend.Base_Source]):
-        self.frame_size: typing.Tuple[int, int] = (
-            int(capture.frame_size[0]),
-            int(capture.frame_size[1]),
-        )
-        self.intrinsics: Camera_Model = capture.intrinsics
-        try:
-            self.timestamps: np.ndarray = capture.timestamps
-        except AttributeError:
-            self.timestamps: np.ndarray = np.ndarray([])
-
-
-Color_RGBA = typing.Tuple[float, float, float, float]
-
-Color_RGB = typing.Tuple[int, int, int]
-
-def color_rgb_to_rgba(rgb: Color_RGB) -> Color_RGBA:
-    r, g, b = rgb
-    return r/255, g/255, b/255, 1.0
-
-@enum.unique
-class Segment_Method(enum.Enum):
-    GAZE = "gaze"
-    PUPIL = "pupil"
-
-
-@enum.unique
-class Segment_Class(enum.Enum):
-    FIXATION = "fixation"
-    SACCADE = "saccade"
-    POST_SACCADIC_OSCILLATIONS = "pso"
-    SMOOTH_PURSUIT = "smooth_pursuit"
-
-    @classmethod
-    def from_nslr_class(cls, nslr_class):
-        return {
-            nslr_hmm.FIXATION: Segment_Class.FIXATION,
-            nslr_hmm.SACCADE: Segment_Class.SACCADE,
-            nslr_hmm.PSO: Segment_Class.POST_SACCADIC_OSCILLATIONS,
-            nslr_hmm.SMOOTH_PURSUIT: Segment_Class.SMOOTH_PURSUIT,
-        }[nslr_class]
-
-    @property
-    def color_rgba(self) -> Color_RGBA:
-        return color_rgb_to_rgba(self.color_rgb)
-
-    @property
-    def color_rgb(self) -> Color_RGB:
-        # https://flatuicolors.com/palette/defo
-        grey = (52, 73, 94)  # wet asphalt
-        yellow = (241, 196, 15)  # sun flower
-        green = (39, 174, 96)  # nephritis
-        blue = (41, 128, 185)  # belize hole
-        purple = (142, 68, 173)  # wisteria
-        return {
-            Segment_Class.FIXATION: yellow,
-            Segment_Class.SACCADE: green,
-            Segment_Class.POST_SACCADIC_OSCILLATIONS: blue,
-            Segment_Class.SMOOTH_PURSUIT: purple,
-        }.get(self, grey)
-
-
-
-
-class _Classified_Segment_Abstract_Storage(metaclass=abc.ABCMeta):
-    def __getitem__(self, key):
-        pass
-    def get(self, key, default):
-        pass
-    def to_dict(self) -> dict:
-        pass
-    def to_serialized_dict(self) -> fm.Serialized_Dict:
-        pass
-    def to_msgpack(self) -> MsgPack_Serialized_Segment:
-        pass
-
-
-class _Classified_Segment_Dict_Storage(_Classified_Segment_Abstract_Storage):
-    def __init__(self, python_dict: dict):
-        self._python_dict = python_dict
-    def __getitem__(self, key):
-        return self._python_dict.__getitem__(key)
-    def get(self, key, default):
-        return self._python_dict.get(key, default)
-    def to_dict(self) -> dict:
-        return self._python_dict
-    def to_serialized_dict(self) -> fm.Serialized_Dict:
-        return fm.Serialized_Dict(python_dict=self._python_dict)
-    def to_msgpack(self) -> MsgPack_Serialized_Segment:
-        serialized_dict = fm.Serialized_Dict(python_dict=self._python_dict)
-        return serialized_dict.serialized
-
-
-class _Classified_Segment_Serialized_Dict_Storage(_Classified_Segment_Abstract_Storage):
-    def __init__(self, serialized_dict: fm.Serialized_Dict):
-        self._serialized_dict = serialized_dict
-    def __getitem__(self, key):
-        return self._serialized_dict.__getitem__(key)
-    def get(self, key, default):
-        return self._serialized_dict.get(key, default)
-    def to_dict(self) -> dict:
-        return dict(self._serialized_dict)
-    def to_serialized_dict(self) -> fm.Serialized_Dict:
-        return self._serialized_dict
-    def to_msgpack(self) -> MsgPack_Serialized_Segment:
-        serialized_dict = self._serialized_dict
-        return serialized_dict.serialized
-
-
-class _Classified_Segment_MsgPack_Storage(_Classified_Segment_Serialized_Dict_Storage):
-    def __init__(self, msgpack_bytes: MsgPack_Serialized_Segment):
-        serialized_dict = fm.Serialized_Dict(msgpack_bytes=msgpack_bytes)
-        super().__init__(serialized_dict=serialized_dict)
-
-
-
-
-class Classified_Segment:
-    def from_attrs(
-        id: int,
-        topic: str,
-        use_pupil: bool,
-        segment_data: Gaze_Data,
-        segment_time: Gaze_Time,
-        segment_class: Segment_Class,
-        start_frame_index: int,
-        end_frame_index: int,
-        start_frame_timestamp: float,
-        end_frame_timestamp: float,
-    ) -> "Classified_Segment":
-
-        segment_dict = {
-            "id": id,
-            "topic": topic,
-            "use_pupil": use_pupil,
-            "segment_data": segment_data,
-            "segment_time": segment_time,
-            "segment_class": segment_class.value,
-            "start_frame_index": start_frame_index,
-            "end_frame_index": end_frame_index,
-            "start_frame_timestamp": start_frame_timestamp,
-            "end_frame_timestamp": end_frame_timestamp,
-        }
-
-        segment_dict["confidence"] = float(
-            np.mean([gp["confidence"] for gp in segment_data])
-        )
-
-        norm_pos_2d_points = np.array([gp["norm_pos"] for gp in segment_data])
-        segment_dict["norm_pos"] = np.mean(norm_pos_2d_points, axis=0).tolist()
-
-        if use_pupil:
-            gaze_3d_points = np.array(
-                [gp["gaze_point_3d"] for gp in segment_data], dtype=np.float32
-            )
-            segment_dict["gaze_point_3d"] = np.mean(gaze_3d_points, axis=0).tolist()
-
-        return Classified_Segment.from_dict(segment_dict)
-
-    def __init__(self, storage: typing.Type[_Classified_Segment_Abstract_Storage]):
-        self._storage = storage
-
-    def validate(self):
-        assert self.frame_count > 0
-        assert len(self.segment_data) == len(self.segment_time) == self.frame_count
-        assert self.start_frame_timestamp <= self.end_frame_timestamp
-        assert self.start_frame_timestamp == self.segment_time[0]
-        assert self.end_frame_timestamp == self.segment_time[-1]
-
-    # Public Format
-
-    def to_public_dict(self) -> dict:
-        """
-        Returns a dictionary representation of the segment,
-        in the format suitable for sending over ZMQ,
-        and consumption by clients external to this plugin.
-        """
-
-        def serialize_gaze_data(gaze_data: np.ndarray) -> typing.Tuple[dict]:
-            return tuple(map(dict, gaze_data))
-
-        def serialize_gaze_time(gaze_time: np.ndarray) -> typing.Tuple[float]:
-            return tuple(gaze_time)
-
-        public_dict = {
-            "id": self.id,
-            "topic": self.topic,
-            "method": self.method.value,
-            # "segment_data": serialize_gaze_data(self.segment_data),
-            # "segment_time": serialize_gaze_time(self.segment_time),
-            "segment_class": self.segment_class.value,
-            "start_frame_index": self.start_frame_index,
-            "end_frame_index": self.end_frame_index,
-            "start_frame_timestamp": self.start_frame_timestamp,
-            "end_frame_timestamp": self.end_frame_timestamp,
-        }
-        return public_dict
-
-    # Serialization
-
-    def to_dict(self) -> dict:
-        return self._storage.to_dict()
-
-    def to_serialized_dict(self) -> fm.Serialized_Dict:
-        return self._storage.to_serialized_dict()
-
-    def to_msgpack(self) -> MsgPack_Serialized_Segment:
-        return self._storage.to_msgpack()
-
-    # Deserialization
-
-    def from_dict(segment_dict: dict) -> "Classified_Segment":
-        storage = _Classified_Segment_Dict_Storage(segment_dict)
-        return Classified_Segment(storage)
-
-    def from_serialized_dict(serialized_dict: fm.Serialized_Dict) -> "Classified_Segment":
-        storage = _Classified_Segment_Serialized_Dict_Storage(serialized_dict)
-        return Classified_Segment(storage)
-
-    def from_msgpack(segment_msgpack: MsgPack_Serialized_Segment) -> "Classified_Segment":
-        storage = _Classified_Segment_MsgPack_Storage(segment_msgpack)
-        return Classified_Segment(storage)
-
-    # Stored properties
-
-    @property
-    def id(self) -> int:
-        """..."""
-        return self._storage["id"]
-
-    @property
-    def topic(self) -> str:
-        """..."""
-        return self._storage["topic"]
-
-    @property
-    def use_pupil(self) -> bool:
-        """..."""
-        return self._storage["use_pupil"]
-
-    @property
-    def segment_data(self) -> list:  # FIXME: Annotate with typed list
-        """..."""
-        return self._storage["segment_data"]
-
-    @property
-    def segment_time(self) -> list:  # FIXME: Annotate with typed list
-        """..."""
-        return self._storage["segment_time"]
-
-    @property
-    def segment_class(self) -> Segment_Class:
-        """..."""
-        return Segment_Class(self._storage["segment_class"])
-
-    @property
-    def start_frame_index(self) -> int:
-        """Index of the first segment frame, in the frame buffer."""
-        return self._storage["start_frame_index"]
-
-    @property
-    def end_frame_index(self) -> int:
-        """Index **after** the last segment frame, in the frame buffer."""
-        return self._storage["end_frame_index"]
-
-    @property
-    def start_frame_timestamp(self) -> float:
-        """Timestamp of the first frame, in the frame buffer."""
-        return self._storage["start_frame_timestamp"]
-
-    @property
-    def end_frame_timestamp(self) -> float:
-        """Timestamp of the last frame, in the frame buffer."""
-        return self._storage["end_frame_timestamp"]
-
-    @property
-    def norm_pos(self):
-        """..."""
-        return self._storage["norm_pos"]
-
-    @property
-    def gaze_point_3d(self):
-        """..."""
-        return self._storage.get("gaze_point_3d", None)
-
-    @property
-    def confidence(self):
-        """..."""
-        return self._storage["confidence"]
-
-    # Computed properties
-
-    @property
-    def method(self) -> Segment_Method:
-        """..."""
-        return Segment_Method.PUPIL if self.use_pupil else Segment_Method.GAZE
-
-    @property
-    def timestamp(self):
-        """..."""
-        return self.start_frame_timestamp
-
-    @property
-    def duration(self) -> float:
-        """Duration in ms."""
-        return (self.end_frame_timestamp - self.start_frame_timestamp) * 1000
-
-    @property
-    def frame_count(self) -> int:
-        """..."""
-        return self.end_frame_index - self.start_frame_index
-
-    @property
-    def mid_frame_index(self):
-        """Index of the middle segment frame, in the frame buffer.
-        """
-        return int((self.end_frame_index + self.start_frame_index) // 2)
-
-    @property
-    def mid_frame_timestamp(self) -> float:
-        """Timestamp of the middle frame, in the frame buffer."""
-        return (self.end_frame_timestamp + self.start_frame_timestamp) / 2
-
-    @property
-    def time_range(self) -> Time_Range:
-        return Time_Range(start_time=self.start_frame_timestamp, end_time=self.end_frame_timestamp)
-
-    @property
-    def color_rgb(self) -> Color_RGB:
-        return self.segment_class.color_rgb
-
-    @property
-    def color_rgba(self) -> Color_RGBA:
-        return self.segment_class.color_rgba
-
-    def mean_2d_point_within_world(
-        self, world_frame: typing.Tuple[int, int]
-    ) -> typing.Tuple[int, int]:
-        x, y = self.norm_pos
-        x, y = methods.denormalize((x, y), world_frame, flip_y=True)
-        return int(x), int(y)
-
-    def last_2d_point_within_world(
-        self, world_frame: typing.Tuple[int, int]
-    ) -> typing.Tuple[int, int]:
-        x, y = self.segment_data[-1]["norm_pos"]
-        x, y = methods.denormalize((x, y), world_frame, flip_y=True)
-        return int(x), int(y)
-
-    def draw_on_frame(self, frame):
-        #TODO: Type annotate frame
-        world_frame = (frame.width, frame.height)
-        segment_point = self.mean_2d_point_within_world(world_frame)
-        segment_color = self.color_rgba
-
-        pm.transparent_circle(
-            frame.img, segment_point, radius=25.0, color=segment_color, thickness=3
-        )
-
-        text = str(self.id)
-        text_origin = (segment_point[0] + 30, segment_point[1])
-        text_fg_color = self.color_rgb
-        font_face = cv2.FONT_HERSHEY_DUPLEX
-        font_scale = 0.8
-        font_thickness = 1
-
-        cv2.putText(
-            img=frame.img,
-            text=text,
-            org=text_origin,
-            fontFace=font_face,
-            fontScale=font_scale,
-            color=text_fg_color,
-            thickness=font_thickness,
-        )
-
-    def draw_in_gl_context(self, frame_size: typing.Tuple[int, int], glfont):
-        #TODO: Type annotate glfont
-        segment_point = self.last_2d_point_within_world(frame_size)
-
-        draw_circle(segment_point, radius=48.0, stroke_width=10.0, color=RGBA(*self.color_rgba))
-
-        font_size = 22
-        text = str(self.id)
-        text_origin_x = segment_point[0] + 48.0
-        text_origin_y = segment_point[1]
-        text_fg_color = self.color_rgba
-
-        glfont.set_size(font_size)
-        glfont.set_color_float(text_fg_color)
-        glfont.draw_text(text_origin_x, text_origin_y, text)
-
-
-class Classified_Segment_Factory:
-    __slots__ = "_segment_id"
-
-    def __init__(self, start_id: int = None):
-        if start_id is None:
-            start_id = 0
-        assert isinstance(start_id, int)
-        self._segment_id = start_id
-
-    def create_segment(
-        self, gaze_data, gaze_time, use_pupil, nslr_segment, nslr_segment_class
-    ) -> typing.Optional[Classified_Segment]:
-        id = self._get_id_postfix_increment()
-
-        i_start, i_end = nslr_segment.i
-        segment_data = list(gaze_data[i_start:i_end])
-        segment_time = list(gaze_time[i_start:i_end])
-
-        if len(segment_data) == 0:
-            return None
-
-        segment_class = Segment_Class.from_nslr_class(nslr_segment_class)
-        topic = "nslr_segmentation"
-
-        start_frame_index, end_frame_index = nslr_segment.i  # [i_0, i_1)
-        start_frame_timestamp, end_frame_timestamp = (
-            segment_time[0],
-            segment_time[-1],
-        )  # [t_0, t_1]
-
-        segment = Classified_Segment.from_attrs(
-            id=id,
-            topic=topic,
-            use_pupil=use_pupil,
-            segment_data=segment_data,
-            segment_time=segment_time,
-            segment_class=segment_class,
-            start_frame_index=start_frame_index,
-            end_frame_index=end_frame_index,
-            start_frame_timestamp=start_frame_timestamp,
-            end_frame_timestamp=end_frame_timestamp,
-        )
-
-        try:
-            segment.validate()
-        except:
-            return None
-
-        return segment
-
-    def _get_id_postfix_increment(self) -> int:
-        id = self._segment_id
-        self._segment_id += 1
-        return id
-
-
 Eye_Movement_Generator_Yield = typing.Tuple[
     str, typing.Optional[MsgPack_Serialized_Segment]
 ]
